requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pytest==8.4.1
openpyxl==3.1.2
python-dateutil>=2.8
